useAppHandlersの責務分離/リファクタリング
